requirements.txt
python-dotenv
openai
requests
flask
langchain
langchain-community
langchain-openai
langchain-chroma
langchain-text-splitters
chromadb
pypdf
PyPDF2
ipykernel
uvicorn
prompty 
azure-identity 
azure-cosmos
azure-ai-openai